tests/wpt/web-platform-tests/css/work-in-progress/gadams/cssom/cssnamespacerule-interface.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD//XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>CSSOM - CSSNamespaceRule - Interface</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="http://www.w3.org/TR/cssom-1/#the-cssnamespacerule-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script><![CDATA[

      test(function(){

        assert_true('CSSNamespaceRule' in window);

      }, '\'CSSNamespaceRule\' in window');

      test(function(){

        assert_true(CSSNamespaceRule instanceof Function);

      }, 'CSSNamespaceRule instanceof Function');

      test(function(){

        assert_true('prototype' in CSSNamespaceRule);

      }, '\'prototype\' in CSSNamespaceRule');

      test(function(){

        assert_equals(CSSNamespaceRule.prototype.toString(), '[object CSSNamespaceRulePrototype]');

      }, 'CSSNamespaceRule.prototype.toString() == \'[object CSSNamespaceRulePrototype]\'');

      test(function(){

        assert_true('constructor' in CSSNamespaceRule.prototype);

      }, '\'constructor\' in CSSNamespaceRule.prototype');

      test(function(){

        assert_equals(CSSNamespaceRule.prototype.constructor,CSSNamespaceRule);

      }, 'CSSNamespaceRule.prototype.constructor === CSSNamespaceRule');

    ]]>
    </script>
  </body>
</html>
